+++ pylxd/deprecated/base.py
