+++ pylxd/network.py
@@ -12,7 +12,7 @@
 #    License for the specific language governing permissions and limitations
 #    under the License.
 
-from pylxd import base
+from pylxd.deprecated import base
 
 
 class LXDNetwork(base.LXDBase):
